yeet-installer/requirements.txt
Pillow>=10.0.0
ttkbootstrap>=1.10.1
